// src/app/model/AccessGroup.ts
export class AccessGroup {
    groupId: number;
    groupName: string;
    settings?: boolean;
    users?: boolean;
    usersReadOnly?: boolean;
    groups?: boolean;
    groupsReadOnly?: boolean;
    general?: boolean;
    generalReadOnly?: boolean;
    stock?: boolean;
    newStock?: boolean;
    costPrice?: boolean;
    costPriceReadOnly?: boolean;
    stockDelete?: boolean;
    stockUpdate?: boolean;
    stockInfo?: boolean;
    saleOrder?: boolean;
    allOrders?: boolean;
    orderCost?: boolean;
    orderChangePrice?: boolean;
    orderDelete?: boolean;
    orderUpdate?: boolean;
    newOrder?: boolean;
    orderCurrentUser?: boolean;
    orderOtherUserReadOnly?: boolean;
    orderOtherUserOpen?: boolean;
    orderOtherUserFind?: boolean;
    orderInfo?: boolean;
    orderInvoice?: boolean;
    orderPayment?: boolean;
    orderPickup?: boolean;
    orderPickupReadOnly?: boolean;

    constructor() {
        this.groupId = 0;
        this.groupName = '';
        this.settings = true;
        this.users = true;
        this.usersReadOnly = false;
        this.groups = true;
        this.groupsReadOnly = false;
        this.general = true;
        this.generalReadOnly = false;
        this.stock = true;
        this.newStock = true;
        this.costPrice = true;
        this.costPriceReadOnly = false;
        this.stockDelete = true;
        this.stockUpdate = true;
        this.stockInfo = true;
        this.saleOrder = true;
        this.allOrders = true;
        this.orderCost = true;
        this.orderChangePrice = true;
        this.orderDelete = true;
        this.orderUpdate = true;
        this.newOrder = true;
        this.orderCurrentUser = false;
        this.orderOtherUserReadOnly = false;
        this.orderOtherUserOpen = true;
        this.orderOtherUserFind = true;
        this.orderInfo = true;
        this.orderInvoice = true;
        this.orderPayment = true;
        this.orderPickup = true;
        this.orderPickupReadOnly = false;
    }
}