// src/app/model/User.ts
export class User {
    userId: number;
    firstName: string;
    lastName: string;
    phone: string;
    email: string;
    password: string;
    passReset: boolean;
    jobTitle: string;
    empType: String;
    groupId: number;
    empStatus: boolean;
    loginTimestamp?: string;
    userComment?: string;
}